[revm] rename Machine to interpreter
