+++ crates/revm/src/interpreter/memory.rs
